// src/lib/api/model/offerList.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * API Documentation
 * API for your Django project
 * OpenAPI spec version: v1
 */
import type { ManufacturerList } from './manufacturerList';
import type { OfferListStatus } from './offerListStatus';

export interface OfferList {
  readonly id?: number;
  manufacturer?: ManufacturerList;
  service: number;
  status?: OfferListStatus;
  created_at?: string;
}